+++ cyral/utils.go
@@ -5,19 +5,6 @@ import (
 	"sort"
 )
 
-func formatAttributes(attributes []string) string {
-	if len(attributes) == 0 {
-		return ""
-	}
-	s := fmt.Sprintf(`"%s"`, attributes[0])
-	if len(attributes) > 1 {
-		for _, attribute := range attributes[1:] {
-			s += fmt.Sprintf(`, "%s"`, attribute)
-		}
-	}
-	return s
-}
-
 func urlQuery(kv map[string]string) string {
 	queryStr := "?"
 	for k, v := range kv {
